requirements.txt
Pillow>=8.4.0
django-appconf>=1.0.5
dnspython>=2.3.0
